# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API Sandbox CESFAM",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/prescriptions": {
      "get": {
        "tags": [
          "Recetas"
        ],
        "summary": "List Prescriptions",
        "description": "Lista recetas. status_filter acepta valores separados por coma: ?status_filter=SUBMITTED,RESERVED",
        "operationId": "list_prescriptions_api_v1_prescriptions_get",
        "parameters": [
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          },
          {
            "name": "patientId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Patientid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Recetas"
        ],
        "summary": "Create Prescription",
        "operationId": "create_prescription_api_v1_prescriptions_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrescriptionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prescriptions/queue": {
      "get": {
        "tags": [
          "Recetas"
        ],
        "summary": "Queue",
        "description": "Recetas pendientes de preparación (SUBMITTED + RESERVED).",
        "operationId": "queue_api_v1_prescriptions_queue_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Prescription"
                  },
                  "title": "Response Queue Api V1 Prescriptions Queue Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prescriptions/{prescription_id}": {
      "get": {
        "tags": [
          "Recetas"
        ],
        "summary": "Get Prescription",
        "operationId": "get_prescription_api_v1_prescriptions__prescription_id__get",
        "parameters": [
          {
            "name": "prescription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prescription Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prescriptions/{prescription_id}/prepare": {
      "post": {
        "tags": [
          "Recetas"
        ],
        "summary": "Prepare",
        "description": "Mueve SUBMITTED → READY_FOR_PICKUP si hay stock disponible.",
        "operationId": "prepare_api_v1_prescriptions__prescription_id__prepare_post",
        "parameters": [
          {
            "name": "prescription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prescription Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Prescription"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prescriptions/{prescription_id}/reserve": {
      "post": {
        "tags": [
          "Recetas"
        ],
        "summary": "Reserve",
        "operationId": "reserve_api_v1_prescriptions__prescription_id__reserve_post",
        "parameters": [
          {
            "name": "prescription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prescription Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Prescription"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prescriptions/{prescription_id}/mark-available": {
      "post": {
        "tags": [
          "Recetas"
        ],
        "summary": "Mark Available",
        "description": "RESERVED → READY_FOR_PICKUP (se dispara cuando llega stock).",
        "operationId": "mark_available_api_v1_prescriptions__prescription_id__mark_available_post",
        "parameters": [
          {
            "name": "prescription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prescription Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Prescription"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prescriptions/{prescription_id}/external-purchase": {
      "post": {
        "tags": [
          "Recetas"
        ],
        "summary": "External Purchase",
        "operationId": "external_purchase_api_v1_prescriptions__prescription_id__external_purchase_post",
        "parameters": [
          {
            "name": "prescription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prescription Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalPurchaseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Prescription"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prescriptions/{prescription_id}/cancel": {
      "post": {
        "tags": [
          "Recetas"
        ],
        "summary": "Cancel",
        "operationId": "cancel_api_v1_prescriptions__prescription_id__cancel_post",
        "parameters": [
          {
            "name": "prescription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prescription Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Prescription"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CancelRequest": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "title": "CancelRequest"
      },
...
      "ExternalPurchaseRequest": {
        "properties": {
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "ExternalPurchaseRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Prescription": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "doctorId": {
            "type": "string",
            "title": "Doctorid"
          },
          "patientId": {
            "type": "string",
            "title": "Patientid"
          },
          "emissionDate": {
            "type": "string",
            "format": "date",
            "title": "Emissiondate"
          },
          "pickupDeadline": {
            "type": "string",
            "format": "date",
            "title": "Pickupdeadline"
          },
          "treatmentType": {
            "$ref": "#/components/schemas/TreatmentType"
          },
          "durationDays": {
            "type": "integer",
            "title": "Durationdays"
          },
          "status": {
            "$ref": "#/components/schemas/PrescriptionStatus"
          },
          "nextScheduledDelivery": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextscheduleddelivery"
          }
        },
        "type": "object",
        "required": [
          "id",
          "doctorId",
          "patientId",
          "emissionDate",
          "pickupDeadline",
          "treatmentType",
          "durationDays",
          "status"
        ],
        "title": "Prescription"
      },
      "PrescriptionCreate": {
        "properties": {
          "patientId": {
            "type": "string",
            "title": "Patientid"
          },
          "treatmentType": {
            "$ref": "#/components/schemas/TreatmentType"
          },
          "durationDays": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Durationdays"
          },
          "pickupDeadline": {
            "type": "string",
            "format": "date",
            "title": "Pickupdeadline"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/PrescriptionLine"
            },
            "type": "array",
            "minItems": 1,
            "title": "Lines"
          }
        },
        "type": "object",
        "required": [
          "patientId",
          "treatmentType",
          "durationDays",
          "pickupDeadline",
          "lines"
        ],
        "title": "PrescriptionCreate"
      },
      "PrescriptionDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "doctorId": {
            "type": "string",
            "title": "Doctorid"
          },
          "patientId": {
            "type": "string",
            "title": "Patientid"
          },
          "emissionDate": {
            "type": "string",
            "format": "date",
            "title": "Emissiondate"
          },
          "pickupDeadline": {
            "type": "string",
            "format": "date",
            "title": "Pickupdeadline"
          },
          "treatmentType": {
            "$ref": "#/components/schemas/TreatmentType"
          },
          "durationDays": {
            "type": "integer",
            "title": "Durationdays"
          },
          "status": {
            "$ref": "#/components/schemas/PrescriptionStatus"
          },
          "nextScheduledDelivery": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextscheduleddelivery"
          },
          "patientFullName": {
            "type": "string",
            "title": "Patientfullname"
          },
          "patientRut": {
            "type": "string",
            "title": "Patientrut"
          },
          "doctorFullName": {
            "type": "string",
            "title": "Doctorfullname"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/PrescriptionLine"
            },
            "type": "array",
            "title": "Lines"
          }
        },
        "type": "object",
        "required": [
          "id",
          "doctorId",
          "patientId",
          "emissionDate",
          "pickupDeadline",
          "treatmentType",
          "durationDays",
          "status",
          "patientFullName",
          "patientRut",
          "doctorFullName",
          "lines"
        ],
        "title": "PrescriptionDetail"
      },
      "PrescriptionLine": {
        "properties": {
          "medicationId": {
            "type": "string",
            "title": "Medicationid"
          },
          "dosesPerInterval": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Dosesperinterval"
          },
          "intervalHours": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Intervalhours"
          },
          "doseDescription": {
            "type": "string",
            "title": "Dosedescription"
          },
          "durationDays": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Durationdays"
          },
          "totalQuantity": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Totalquantity"
          }
        },
        "type": "object",
        "required": [
          "medicationId",
          "dosesPerInterval",
          "intervalHours",
          "doseDescription",
          "durationDays",
          "totalQuantity"
        ],
        "title": "PrescriptionLine"
      },
      "PrescriptionStatus": {
        "type": "string",
        "enum": [
          "SUBMITTED",
          "RESERVED",
          "READY_FOR_PICKUP",
          "PICKED_UP",
          "EXTERNAL_PURCHASE",
          "CANCELLED",
          "EXPIRED"
        ],
        "title": "PrescriptionStatus"
      },
...
      "TreatmentType": {
        "type": "string",
        "enum": [
          "SHORT",
          "LONG"
        ],
        "title": "TreatmentType"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}